Update README.md
Introduction to if statements

Step 1

Edit the program so that it works properly for the correct legal driving age. Test that it works properly by changing the value of the age variable.

Then, modify the program so that it prompts the user to enter how old they are by typing in a number from the keyboard.

Sanple run:

Please enter your age 16
You are old enough to drive

Step 2

Add an else block to the if statement that states the user is not old enough to drive , creating an if-else statement. The bottom of your code should look like this:

   cout << "You are old enough to drive\n";
  else
   cout << "You are NOT old enough to drive\n";
  }
}

Test your code and make sure it works properly.